+++ scripts/validate-template-parameters.sh
@@ -14,7 +14,7 @@ validate_env_var "$APP_NAME" appName
 validate_env_var "$APP_TEST" appTest
 validate_env_var "$AZURE_SUBSCRIPTION" azureSubscription
 validate_env_var "$ENVIRONMENT" environment
-validate_env_var "$RELEASE_ENVIRONMENT" releaseEnvironment
+validate_env_var "$VAULT_RELEASE_ENVIRONMENT" releaseEnvironment
 
 if [ "$FAIL" = 1 ]; then
   echo "##vso[task.complete result=Failed;]"