+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.2.0 / TBC
 ===================
 - Treat empty object entries as erroneous
+- Add information about deployment environment variables
 
 0.1.0 / 2019-10-28
 ===================